requirements/local-base.txt
hydra-core>=1.3,<1.4
omegaconf>=2.3,<2.4
python-dotenv>=1.0
numpy<2
pandas>=2.0,<2.3
scipy>=1.10,<1.14
sympy>=1.12,<1.14
networkx>=3.1,<3.4
scikit-learn>=1.3,<1.6
ase>=3.22,<3.24
pymatgen>=2024.2,<2025.0
matminer>=0.9,<1.0
tqdm>=4.66,<5
p-tqdm>=1.4,<2
smact>=2.5,<3
wandb>=0.16,<0.19
pytest>=8,<9
